Move Compute-Only Expressions from Air=>Mir
This separates computation from expression evaluation.  More
specifically, there are no `air.Expr` forms which cannot be evaluated by
the prover.  Instead, `ComputedColumn` now accepts an _arbitrary_
computation which can be evaluated, rather than (for example) an
instance of `air.Expr`.

The net effect is that `air.Inverse` is now moved up to `mir.Inverse`.
